pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-sc2-remake"
version = "0.1.0"
description = "Clean-room remake of a StarCraft II match outcome prediction project"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Jose Luis Pineda Fritas"}
]
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "scikit-learn",
  "matplotlib",
  "PyYAML",
  "tqdm",
  "joblib",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
